move Icons/checkIcon to Form/CheckIcon
